+++ Trace.ts
@@ -1,100 +1,96 @@
 import { ExitStatus } from "./common.ts";
-import { assert } from "./deps/std/testing/asserts.ts";
 import { AnyEffect } from "./Effect.ts";
 import { Hooks } from "./Hooks.ts";
+import { Work } from "./Work.ts";
 
-export type TraceEventDesc<
-  EnterResult,
-  ExitResult_,
-> = TraceEnterEventDec<EnterResult> | TraceExitEventDesc<ExitResult_>;
-
-export type TraceEventHookDesc = TraceEventDesc<unknown, ExitStatus>;
-
-export interface Trace extends Hooks {
-  digest: () => TraceEventHookDesc[];
+export function trace(): Trace {
+  return new Trace();
 }
+export class Trace extends Hooks {
+  elements: TraceElement[] = [];
 
-export function trace(...events: TraceEventHookDesc[]): Trace {
-  let sequence = [...events];
-  return {
-    digest: () => {
-      const s = sequence;
-      sequence = [];
-      return s;
-    },
-    enter: (work, result) => {
-      sequence.push(new TraceEnterEventDec(work.source, result));
-    },
-    exit: (work, result) => {
-      sequence.push(new TraceExitEventDesc(work.source, result));
-    },
+  enter = (work: Work, result: unknown): void => {
+    this.elements.push({
+      kind: "enter",
+      source: work.source,
+      result,
+    });
+  };
+
+  exit = (work: Work, result: ExitStatus): void => {
+    this.elements.push({
+      kind: "exit",
+      source: work.source,
+      result,
+    });
   };
 }
 
-export function traceDesc(...initial: TraceAssertions[]) {
-  return new TraceDescBuilder(initial);
+export type TraceElement =
+  | TraceElementBase<"enter", unknown>
+  | TraceElementBase<"exit", ExitStatus>;
+type TraceElementBase<Kind extends "enter" | "exit", Result> = {
+  kind: Kind;
+  source: AnyEffect;
+  result: Result;
+};
+
+export const traceMockElements_ = Symbol();
+
+export function traceMock(...traceElement: TraceMockElement[]): TraceMock {
+  return new TraceMock(traceElement);
 }
+export class TraceMock {
+  [traceMockElements_]: TraceMockElement[];
 
-export type TraceAssertions = TraceEventDesc<
-  (useResult?: unknown) => void,
-  (useResult?: ExitStatus) => void
->;
-
-export class TraceDescBuilder {
-  constructor(readonly assertions: TraceAssertions[]) {}
-
-  digest = (): TraceAssertions[] => {
-    return [...this.assertions];
-  };
+  constructor(elements: TraceMockElement[] = []) {
+    this[traceMockElements_] = elements;
+  }
 
   enter = (
     source: AnyEffect,
-    useResult?: (result?: unknown) => void,
-  ) => {
-    return new TraceDescBuilder([
-      ...this.assertions,
-      new TraceEnterEventDec(source, useResult),
-    ]);
+    result?: (result: unknown) => void,
+  ): TraceMock => {
+    return new TraceMock([...this[traceMockElements_], {
+      kind: "enter",
+      result,
+      source,
+    }]);
   };
 
   exit = (
     source: AnyEffect,
-    useResult?: (result?: ExitStatus) => void,
-  ) => {
-    return new TraceDescBuilder([
-      ...this.assertions,
-      new TraceExitEventDesc(source, useResult),
-    ]);
+    result?: (result: ExitStatus) => void,
+  ): TraceMock => {
+    return new TraceMock([...this[traceMockElements_], {
+      kind: "exit",
+      result,
+      source,
+    }]);
+  };
+
+  branch = (
+    ...branches: ((builder: Omit<TraceMock, "branch">) => TraceMock)[]
+  ): TraceMock => {
+    return new TraceMock([...this[traceMockElements_], {
+      kind: "branching",
+      branches: branches.map((branch) => branch(new TraceMock())),
+    }]);
   };
 }
 
-export class TraceEnterEventDec<Result> {
-  readonly kind = "enter";
-  constructor(
-    readonly source: AnyEffect,
-    readonly result?: Result,
-  ) {}
-}
-
-export class TraceExitEventDesc<Result> {
-  readonly kind = "exit";
-  constructor(
-    readonly source: AnyEffect,
-    readonly result?: Result,
-  ) {}
-}
+export type TraceMockElement =
+  | TraceElementBase<"enter", void | ((result: unknown) => void)>
+  | TraceElementBase<"exit", void | ((result: ExitStatus) => void)>
+  | {
+    kind: "branching";
+    branches: TraceMock[];
+  };
 
 export function assertTrace(
-  trace: Trace,
-  traceAssertionBuilder: TraceDescBuilder,
-) {
-  const digest = trace.digest();
-  assert(digest.length === traceAssertionBuilder.assertions.length);
-  digest.forEach((desc, i) => {
-    const expected = traceAssertionBuilder.assertions[i];
-    assert(expected);
-    assert(desc.kind === expected.kind);
-    assert(desc.source === expected.source);
-    expected.result?.(desc.result as any);
-  });
+  _tracer: Trace,
+  mock: TraceMock,
+): void {
+  const _mockElements = mock[traceMockElements_];
+  // TODO
 }
